Make install script generic

+++ install.sh
@@ -1,8 +1,22 @@
 #!/bin/bash
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
-for file in .bashrc
+cd "$DIR"
+find . -type f -not -path "./.git/*" -not -path "./install.sh" | while read file
 do
-	rm -v "$HOME/$file"
-	ln -vs "$DIR/$file" "$HOME/$file"
+    d=$(dirname "$file")
+    if [ ! -d "$HOME/$d" ]
+    then
+        mkdir -pv "$HOME/$d"
+    fi
+    if [ -e "$HOME/$file" ]
+    then
+        diff -uw "$HOME/$file" "$DIR/$file"
+        if [ $? -eq 1 ]
+        then
+            rm -v "$HOME/$file"
+            ln -vs "$DIR/$file" "$HOME/$file"
+        fi
+    else
+        ln -vs "$DIR/$file" "$HOME/$file"
+    fi
 done
-
